test: decouple ios talk parsing coverage
(cherry picked from commit da3cccb21233e2298cc693ef0071dd5a687345c0)
